Silence -Wbad-function-cast warning (when DEBUG_SMARTLIST is on)

+++ src/or/routerlist.c
@@ -2764,7 +2764,7 @@ frac_nodes_with_descriptors(const smartlist_t *sl,
       if (node_has_any_descriptor(node))
         n_with_descs++;
     });
-    return ((double)n_with_descs) / (double)smartlist_len(sl);
+    return ((double)n_with_descs) / smartlist_len(sl);
   }
 
   present = 0.0;
